GPT/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Flask-Migrate==4.0.5
boto3==1.34.54
botocore==1.34.54
jmespath==1.0.1
psycopg2-binary==2.9.9
async_timeout==4.0.3